feat(website): extract shared CHANGELOG parser, auto-generate About journey
parseChangelog.ts extracts the parser from changelog/page.tsx into a shared
lib. app/about/page.tsx becomes a server component that reads CHANGELOG.md
and auto-generates the Project Journey timeline — strips markdown from bullet
text, oldest-first, prepended with 3 narrative milestones.

TimelineClient.tsx holds the IntersectionObserver + Framer Motion animation
logic split from the now-server about page.
